+++ collector-bitmovin-player/src/test/java/com/bitmovin/analytics/bitmovin/player/IBitmovinPlayerCollectorKotlinTest.kt
@@ -1,7 +1,9 @@
 package com.bitmovin.analytics.bitmovin.player
 
 import com.bitmovin.analytics.BitmovinAnalyticsConfig
+import com.bitmovin.analytics.api.AnalyticsConfig
 import io.mockk.mockk
+import org.junit.Assert
 import org.junit.Test
 
 class IBitmovinPlayerCollectorKotlinTest {
@@ -15,5 +17,16 @@ class IBitmovinPlayerCollectorKotlinTest {
         // act and assert (since return value is non nullable we get a valid object or an exception with this call)
         val collector = IBitmovinPlayerCollector.Factory.create(config, mockk(relaxed = true))
         val collector2 = IBitmovinPlayerCollector.create(config, mockk(relaxed = true))
+    }
+
+    // This test is a sanity check that the kotlin factory stays stable and consistent with the naming
+    @Test
+    fun testFactory_shouldCreateNewCollectorObjectWithAnalyticsConfig() {
+        val analyticsConfig = AnalyticsConfig("analytics_key", "player_key")
+        val collector = IBitmovinPlayerCollector.Factory.create(analyticsConfig, mockk(relaxed = true))
+        Assert.assertEquals(analyticsConfig.key, collector.config.key)
+
+        val collector2 = IBitmovinPlayerCollector.create(analyticsConfig, mockk(relaxed = true))
+        Assert.assertEquals(analyticsConfig.key, collector2.config.key)
     }
 }